Nov 3 Product/Category CRUD + Starting Swagger

+++ Back-End/CapstoneOctVGC-version3/src/main/java/com/capstone/controller/CategoryController.java
@@ -2,7 +2,6 @@ package com.capstone.controller;
 
 import java.util.List;
 import java.util.NoSuchElementException;
-import java.util.Objects;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,9 +10,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -83,53 +79,15 @@ public class CategoryController {
 		}
 	}
 	
-	@PostMapping("add_to_product/{prdId}")
-	public ResponseEntity<Category> addCategory(@PathVariable Integer prdId, @RequestBody Category catRequest){
-		Category categ = prdRepo.findById(prdId).map(product -> {
-			int catId = catRequest.getId();
-			if (catId!=0) {
-				Category category = catService.findByCategoryId(catId)
-						.orElseThrow(()-> new NoSuchElementException("Category with id " +catId+" not found."));
-				product.addCategory(category);
-				prdRepo.save(product);
-				return category;
-			}
-			product.addCategory(catRequest);
-			return catRepo.save(catRequest);
-		}).orElseThrow(()-> new NoSuchElementException("Product with id " +prdId+" not found."));
-		return new ResponseEntity<>(categ, HttpStatus.OK);
-	}
-	
-	@PutMapping("/update/{id}")
-	public ResponseEntity<Category> updateCategory(@PathVariable Integer id, @RequestBody Category cat){
-		try {
-			Category databaseCat = catService.findByCategoryId(id).get();
-			if(Objects.nonNull(cat.getCategoryName())) {
-				databaseCat.setCategoryName(cat.getCategoryName());
-			} 
-			return new ResponseEntity<>(catService.save(databaseCat), HttpStatus.OK);
-		}catch (NoSuchElementException e) {
-			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-		}
-	}
-	
-	@DeleteMapping("/remove/{tagId}/from_product/{prdId}")
-	public ResponseEntity<HttpStatus> deleteCategoryFromProduct(@PathVariable Integer tagId, @PathVariable Integer prdId){
-		try {
-			Product product = prdService.findByProductId(prdId).get();
-			product.removeCategory(tagId);
-			prdService.save(product);
-			return new ResponseEntity<>(HttpStatus.ACCEPTED);
-		} catch (NoSuchElementException e) {
-			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-		}
-	}
-	
 	@DeleteMapping("remove/{id}")
 	public ResponseEntity<HttpStatus> deleteCategoryById(@PathVariable Integer id){
 		try {
-			catService.deleteCategoryById(id);
-			return new ResponseEntity<>(HttpStatus.ACCEPTED);
+			if (prdRepo.findProductsByCategoriesId(id)!=null) {
+				return new ResponseEntity<>(HttpStatus.CONFLICT);
+			} else {
+				catService.deleteCategoryById(id);
+				return new ResponseEntity<>(HttpStatus.ACCEPTED);
+			}
 		} catch (NoSuchElementException e) {
 			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
 		}
